% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFF.R
\name{FFF}
\alias{FFF}
\title{Friends and Foe and Famous function}
\usage{
FFF(Year, Month, Day)
}
\arguments{
\item{Year}{A number in "yyyy" format}

\item{Month}{A number in "mm" format}

\item{Day}{A number in "dd" format}
}
\value{
a dataframe of related friends, foe, and famous person of the birthday input
}
\description{
This is the function that get the friends, foe, and famous person  
related to the birthday desired from birthday analyzer.
}
\details{
This function asks the user to input birth year, month and day. 
The function will return the related friends, foe, and famous  
person on the birthday of this user's input.
}
\examples{
Horoscope(1993,11,15)
Horoscope(1962,09,14)
}
