% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNelement.R
\name{FNelement}
\alias{FNelement}
\title{Element function}
\usage{
FNelement(url)
}
\arguments{
\item{url}{from birthday analyzer}
}
\value{
a dataframe of related element the birthday url input
}
\description{
This is the function that get the element of the birthday desired from birthday analyzer web url.
}
\details{
This function asks the user to input url of the birthday analyzer page. 
The function will return the related element of the birthday of this user's input.
}
