% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNfamous.R
\name{FNfamous}
\alias{FNfamous}
\title{Famous function}
\usage{
FNfamous(url)
}
\arguments{
\item{url}{from birthday analyzer}
}
\value{
a dataframe of related 3 famous people born on the same year of the birthday url input
}
\description{
This is the function that get 3 famous people born on the same year of the birthday desired from birthday analyzer web url.
}
\details{
This function asks the user to input url of the birthday analyzer page. 
The function will return the related 3 famous people born on the same year of the birthday of this user's input.
}
