% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNfoes.R
\name{FNfoes}
\alias{FNfoes}
\title{Foes function}
\usage{
FNfoes(url)
}
\arguments{
\item{url}{from birthday analyzer}
}
\value{
a dataframe of related 2 starsigns that are not compatible with the birthday url input
}
\description{
This is the function that get 2 starsigns that are not compatible with the birthday desired from birthday analyzer web url.
}
\details{
This function asks the user to input url of the birthday analyzer page. 
The function will return the related 2 starsigns that are not compatible with the birthday of this user's input.
}
