% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNstarsign.R
\name{FNstarsign}
\alias{FNstarsign}
\title{Starsign function}
\usage{
FNstarsign(url)
}
\arguments{
\item{url}{from birthday analyzer}
}
\value{
a dataframe of related starsign the birthday url input
}
\description{
This is the function that get the starsign of the birthday desired from birthday analyzer web url.
}
\details{
This function asks the user to input url of the birthday analyzer page. 
The function will return the related starsign of the birthday of this user's input.
}
