% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNstone.R
\name{FNstone}
\alias{FNstone}
\title{Stone function}
\usage{
FNstone(url)
}
\arguments{
\item{url}{from birthday analyzer}
}
\value{
a dataframe of related birthstone the birthday url input
}
\description{
This is the function that get the birthstone of the birthday desired from birthday analyzer web url.
}
\details{
This function asks the user to input url of the birthday analyzer page. 
The function will return the related birthstone of the birthday of this user's input.
}
