% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Url.R
\name{Get_Url}
\alias{Get_Url}
\title{Get_Url function}
\usage{
Get_Url(Year = 1993, Month = 11, Day = 15)
}
\arguments{
\item{Year}{A number in "yyyy" format}

\item{Month}{A number in "mm" format}

\item{Day}{A number in "dd" format}
}
\value{
website url of the birthday inout
}
\description{
This is the function that get the webstie of the birthday desired for birthday analyzer.
}
\details{
This function asks the user to input birth year, month and day. 
The function will return the website link on one birthday of 
this user's input.
}
\examples{
Get_Url(1993,11,15)
Get_Url(1962,09,14)
}
