% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horoscope.R
\name{Horoscope}
\alias{Horoscope}
\title{Horoscope function}
\usage{
Horoscope(Year, Month, Day)
}
\arguments{
\item{Year}{A number in "yyyy" format}

\item{Month}{A number in "mm" format}

\item{Day}{A number in "dd" format}
}
\value{
a dataframe of related starsign, element birthstone of the birthday input
}
\description{
This is the function that get the starsign, element, and birthstone
of the birthday desired from birthday analyzer.
}
\details{
This function asks the user to input birth year, month and day. 
The function will return the related starsign, element, and 
birthstone on rhe birthday of this user's input.
}
\examples{
Horoscope(1993,11,15)
Horoscope(1962,09,14)
}
