% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\name{sampledata}
\alias{sampledata}
\title{Create sample dataset function}
\usage{
sampledata(begin, end, frequency)
}
\arguments{
\item{begin}{A date in "yyyy-mm-dd" format}

\item{end}{A date in "yyyy-mm-dd" format}

\item{frequency}{A frequency of months or days}
}
\value{
a saved dataset from begin to end, one birthday out of every month's starsign, guarding element, birthstone, 3 friends, 2 foes, and 3 famous people of that year.
}
\description{
This is the code that creates a dataset of birthday analyzer.
}
\details{
This function takes the user to input two dates, 
first is the begin date and second is the end date. 
The function will return the data on one birthday of 
each month between the begin date and end date.
}
\examples{
sampledata("1990-01-01","1992-01-01","months")
sampledata("1990-01-01","1992-01-01","days")
}
